Fix Zenn data fetching: use server-side rendering like /blog and /products
- Convert app/page.tsx to server component (async function)
- Fetch Zenn articles and products server-side at build time
- Create HomePageClient component to handle client-side scroll logic
- Pass fetched data as props to BlogSection and ProductsSection
- Add revalidate: 3600 for ISR (same as /blog and /products pages)
- Remove client-side data fetching that was causing empty state
- Now uses same proven approach as working /blog page
